+++ libs/core-wrappers/src/index.ts
@@ -5,11 +5,11 @@ export { default as SubmitWrapper } from './lib/SubmitWrapper';
 export { default as Wizard, useWizard, WizardNavigation } from './lib/Wizard';
 export type WizardStep<T> = WizardStepType<T>;
 export {
-  FieldWrapperFormik,
+  FormikFieldWrapper,
   FormikInput,
   FormikSelect,
   FormikCheck,
   FormikTextarea,
   FormikFileInput,
-} from './lib/FieldWrapperFormik';
+} from './lib/FormikFieldWrapper';
 export { default as withBuilder } from './lib/utils/withBuilder';
